# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RV Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get all RV listings
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: state
          in: query
          schema:
            type: string
        - name: make
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: number
        - name: minYear
          in: query
          schema:
            type: number
        - name: maxYear
          in: query
          schema:
            type: number
        - name: maxLength
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: dealScore
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a single RV listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Listing details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    DealScore:
      type: string
      enum:
        - great_deal
        - good_deal
        - fair_deal
        - high_price
        - overpriced
    RvType:
      type: string
      enum:
        - class_a
        - class_b
        - class_c
        - travel_trailer
        - fifth_wheel
        - toy_hauler
        - popup_camper
        - truck_camper
    Listing:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: number
        type:
          $ref: "#/components/schemas/RvType"
        price:
          type: number
        marketValue:
          type: number
        dealScore:
          $ref: "#/components/schemas/DealScore"
        dealSavings:
          type: number
        mileage:
          type: number
        length:
          type: number
        slides:
          type: number
        location:
          type: string
        state:
          type: string
        dealerName:
          type: string
        dealerId:
          type: number
        images:
          type: array
          items:
            type: string
        daysOnMarket:
          type: number
        sleeps:
          type: number
        condition:
          type: string
        isNew:
          type: boolean
        isFeatured:
          type: boolean
        vin:
          type: string
        pricePerFoot:
          type: number
      required:
        - id
        - title
        - make
        - model
        - year
        - type
        - price
        - marketValue
        - dealScore
        - location
        - state
        - dealerName
        - dealerId
        - images
        - daysOnMarket
        - sleeps
        - condition
        - isNew
    ListingDetail:
      allOf:
        - $ref: "#/components/schemas/Listing"
        - type: object
          properties:
            description:
              type: string
            features:
              type: array
              items:
                type: string
            dryWeight:
              type: number
            gvwr:
              type: number
            hitchWeight:
              type: number
            freshWater:
              type: number
            greyWater:
              type: number
            blackWater:
              type: number
            generator:
              type: boolean
            solar:
              type: boolean
            awning:
              type: boolean
            outdoor_kitchen:
              type: boolean
            washerDryer:
              type: boolean
            priceHistory:
              type: array
              items:
                $ref: "#/components/schemas/PricePoint"
            dealer:
              $ref: "#/components/schemas/DealerInfo"
            similarListings:
              type: array
              items:
                $ref: "#/components/schemas/Listing"
    PricePoint:
      type: object
      properties:
        date:
          type: string
        price:
          type: number
      required:
        - date
        - price
    DealerInfo:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        city:
          type: string
        state:
          type: string
        phone:
          type: string
        rating:
          type: number
        reviewCount:
          type: number
        avgResponseTime:
          type: string
        beginnerFriendly:
          type: boolean
        yearsInBusiness:
          type: number
        totalListings:
          type: number
      required:
        - id
        - name
        - city
        - state
        - rating
        - reviewCount
        - avgResponseTime
        - beginnerFriendly
        - yearsInBusiness
        - totalListings
    ListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: number
        offset:
          type: number
        limit:
          type: number
      required:
        - listings
        - total
        - offset
        - limit
...